backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
supabase==2.15.2
python-dotenv==1.1.0
python-multipart==0.0.20
Pillow==11.2.1
pydantic-settings==2.9.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.4.0
google-auth==2.38.0
httpx==0.27.2
